Added registration and some extra features.

+++ admin/includes/view_all_posts.php
@@ -91,7 +91,7 @@
                 <a href='../post.php?p_id={$id}'>View Post</a></td>";
                 echo "<td>
                 <a href='posts.php?source=edit_post&p_id={$id}'>Edit</a><br>
-                <a href='posts.php?delete={$id}'>Delete</a></td>";
+                <a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='posts.php?delete={$id}'>Delete</a></td>";
                 echo "</tr>";
             } ?>
     </tbody>
